requirements.txt
streamlit
google.generativeai
python-dotenv
pymongo
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
pillow
pandas
csv
tempfile
subprocess